requirements.txt
asgiref==3.5.2
certifi==2022.6.15.2
charset-normalizer==2.1.1
dj-static==0.0.6
Django==3.2.15
django-appconf==1.0.5
django-compressor==4.1
django-csp==3.6
django-dsfr==0.9.0
django-referrer-policy==1.0
django-sass-processor==1.2.1
django-widget-tweaks==1.4.12
gunicorn==20.1.0
idna==3.4
libsass==0.21.0
psycopg2-binary==2.9.3
python-dotenv==0.21.0
pytz==2022.2.1
sqlparse==0.4.2
flake8==5.0.4
black==22.8.0
isort==5.10.1
pre-commit==2.20.0
urllib3==1.26.12
requests==2.28.1
sib-api-v3-sdk==7.5.0
factory-boy==3.2.1
behave==1.2.6
selenium==4.4.3
behave-django==1.4.0
django-debug-toolbar==3.6.0
django_rq==2.5.1